Implement default grid
I added a container to index.html for the grid to be created within.

In script.js I implemented a function that would allow me to create a grid
of a given number of divs using DOM methods.

In style.css I added flexbox to each div to place it in the desired location
on the page.